Fix bugs in implementation
It now handles tables inside of a table in an array of tables
correctly.
It now handles arrays of tables inside of a table in an array of
tables correctly.
It will definitely print stuff which should appear at the top level of
a table in an array of tables first.