Fix #17 with comma inside string in array
The array is split on commas and then each result of the split is
parsed as a value. However, with commas inside the string, string
fragments were being passed to the value parser. This change detects
if the array is an array of strings and if so, it looks for string
fragments and concatenates them.