Update Flake8 Preferences
Enable all the errors - the code isn't too bad stylistically.

Remove maximum complexity threshold. The code base is rather complex
(for now).

Set maximum line length to 80. This is slightly more generous than the
default value of 79 but less generous than the previously suggested
120.

Allow certain lines to be ignored - this is necessary for lines which
will trip up the checks in python2 but not python3 or vice versa due
to differences between the python versions.