Add trove classifier for license
The trove classifiers are listed on PyPI to help users know -- at a
glance -- what license the project uses. Helps users decide if the
library is appropriate for integration. A full list of available trove
classifiers can be found at:

https://pypi.org/pypi?%3Aaction=list_classifiers

The setuptools "license" argument is not intended to use trove
classifier notation. Simplify it to "MIT". Details can be found:

https://docs.python.org/3/distutils/setupscript.html#additional-meta-data